Implement `construct_mock_runtime`
This simplify mock creation and provide a default Adapther that can be re-used easilly.
