+++ frame/asset-registry/Cargo.toml
@@ -19,6 +19,7 @@ frame-system = { default-features = false, git = "https://github.com/tidelabs/su
 pallet-sudo = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-balances = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-assets = { default-features = false, path = "../assets" }
+pallet-timestamp = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 sp-runtime = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 sp-core = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 frame-benchmarking = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain", optional = true }
@@ -26,6 +27,7 @@ sp-std = {  default-features = false, git = "https://github.com/tidelabs/substra
 tidefi-primitives = { git = "https://github.com/tidelabs/primitives", branch = "dev", default-features = false }
 serde      = { version = '1', features = ['derive'], optional = true }
 scale-info = { version = "2.0", default-features = false }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -43,8 +45,10 @@ std = [
     'pallet-balances/std',
     'pallet-sudo/std',
     'pallet-assets/std',
+    "pallet-timestamp/std",
     'tidefi-primitives/std',
     "serde",
     "frame-benchmarking/std",
+    "frame-utils/std"
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-support/runtime-benchmarks"]
