+++ frame/fees/Cargo.toml
@@ -30,6 +30,7 @@ pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-tidefi-stake = { path = "../tidefi-stake", default-features = false }
 sp-arithmetic = {  default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 frame-benchmarking = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain", optional = true }
+frame-utils = { path = "../utils", default-features = false }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -53,5 +54,6 @@ std = [
     "pallet-asset-registry/std",
     "pallet-tidefi-stake/std",
     "pallet-sunrise/std",
+    "frame-utils/std"
 ]
 runtime-benchmarks = ["frame-benchmarking"]
