+++ frame/oracle/Cargo.toml
@@ -31,6 +31,7 @@ pallet-fees = { path = "../fees", default-features = false }
 pallet-sunrise = { path = "../sunrise", default-features = false }
 pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-tidefi-stake = { path = "../tidefi-stake", default-features = false }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -56,6 +57,7 @@ std = [
     "pallet-asset-registry/std",
     "pallet-tidefi-stake/std",
     "pallet-sunrise/std",
+    "frame-utils/std"
 
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-system/runtime-benchmarks"]
