+++ frame/quorum/Cargo.toml
@@ -19,6 +19,7 @@ frame-support = { default-features = false, git = "https://github.com/tidelabs/s
 frame-system = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-sudo = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-balances = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
+pallet-timestamp = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-assets = { default-features = false, path = "../assets" }
 pallet-sunrise = { path = "../sunrise", default-features = false }
 sp-runtime = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
@@ -29,6 +30,7 @@ serde      = { version = '1', features = ['derive'], optional = true }
 pallet-security = { path = "../security", default-features = false }
 pallet-asset-registry = { path = "../asset-registry", default-features = false }
 rand_chacha = { version = "0.3.1", default-features = false }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -45,11 +47,13 @@ std = [
     'sp-runtime/std',
     'pallet-balances/std',
     'pallet-sudo/std',
+    'pallet-timestamp/std',
     'tidefi-primitives/std',
     "serde",
     "frame-benchmarking/std",
     "pallet-security/std",
     "pallet-asset-registry/std",
     "pallet-sunrise/std",
+    "frame-utils/std"
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-system/runtime-benchmarks"]
