+++ frame/security/Cargo.toml
@@ -19,6 +19,7 @@ frame-support = { default-features = false, git = "https://github.com/tidelabs/s
 frame-system = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-sudo = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-balances = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
+pallet-timestamp = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-assets = { default-features = false, path = "../assets" }
 sp-runtime = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 sp-core = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
@@ -27,6 +28,7 @@ sp-std = {  default-features = false, git = "https://github.com/tidelabs/substra
 tidefi-primitives = { git = "https://github.com/tidelabs/primitives", branch = "dev", default-features = false }
 serde      = { version = '1', features = ['derive'], optional = true }
 sha2 = { version = "0.10.2", default-features = false } 
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -42,10 +44,12 @@ std = [
     'frame-system/std',
     'sp-runtime/std',
     'pallet-balances/std',
+    'pallet-timestamp/std',
     'pallet-sudo/std',
     'tidefi-primitives/std',
     "serde",
     "frame-benchmarking/std",
     "sha2/std",
+    "frame-utils/std"
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-system/runtime-benchmarks"]
