+++ frame/sunrise/Cargo.toml
@@ -28,6 +28,7 @@ serde      = { version = '1', features = ['derive'], optional = true }
 pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-tidefi-stake = { path = "../tidefi-stake", default-features = false }
 sp-arithmetic = {  default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -49,4 +50,5 @@ std = [
     "pallet-security/std",
     "pallet-asset-registry/std",
     "pallet-tidefi-stake/std",
+    "frame-utils/std"
 ]
