+++ frame/tidefi-stake/Cargo.toml
@@ -31,6 +31,7 @@ pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-timestamp = { default-features = false, git = "https://github.com/tidelabs/substrate", branch = "tidechain" }
 pallet-security = { default-features = false, path = "../security" }
 log = { version = "0.4.17", default-features = false }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -63,7 +64,8 @@ std = [
     "pallet-oracle/std",
     "pallet-fees/std",
     "pallet-sunrise/std",
-    "pallet-tidefi/std"
+    "pallet-tidefi/std",
+    "frame-utils/std"
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-system/runtime-benchmarks"]
 try-runtime = ["frame-support/try-runtime"]
