+++ frame/tidefi/Cargo.toml
@@ -34,6 +34,7 @@ pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-fees = { path = "../fees", default-features = false }
 pallet-sunrise = { path = "../sunrise", default-features = false }
 pallet-tidefi-stake = { path = "../tidefi-stake", default-features = false }
+frame-utils = { default-features = false, path = "../utils" }
 
 [dev-dependencies]
 serde = { version = "1.0.119" }
@@ -61,5 +62,6 @@ std = [
     "pallet-timestamp/std",
     "pallet-tidefi-stake/std",
     "pallet-sunrise/std",
+    "frame-utils/std"
 ]
-runtime-benchmarks = ["frame-benchmarking"]
+runtime-benchmarks = ["frame-benchmarking", "frame-system/runtime-benchmarks"]
