+++ frame/utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+version = "0.6.4"
+name = "frame-utils"
+authors = ["Semantic Network Team <publishers@tidelabs.org>"]
+edition = "2021"
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[features]
+default = ['std']
+std = []
